# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Desafio Técnico Sicredi",
        "description": "Aplicação de Votação",
        "contact": "Adriano Anschau <adrianoanschau@gmail.com>",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/votes": {
            "get": {
                "tags": [
                    "Listar"
                ],
                "summary": "Listar votos de uma Pauta",
                "operationId": "voteList",
                "parameters": [
                    {
                        "name": "schedule_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Pauta não encontrado",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/votes/result": {
            "get": {
                "tags": [
                    "Exibir"
                ],
                "summary": "Exibe contagem de votos de uma Pauta",
                "operationId": "voteResult",
                "parameters": [
                    {
                        "name": "schedule_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Pauta não Encontrada",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    }
}